+++ packages/zowe-explorer/__tests__/__unit__/dataset/dsNodeActions.unit.test.ts
@@ -209,7 +209,7 @@ describe("dsNodeActions", () => {
                 }),
             });
             const spy = jest.spyOn(refreshActions, "refreshAll");
-            refreshActions.refreshAll(testDSTree);
+            await refreshActions.refreshAll(testDSTree);
             expect(spy).toHaveBeenCalledTimes(1);
         });
     });