+++ packages/zowe-explorer/src/uss/init.ts
@@ -66,11 +66,11 @@ export async function initUSSProvider(context: vscode.ExtensionContext): Promise
         vscode.commands.registerCommand("zowe.uss.refreshUSSInTree", (node: IZoweUSSTreeNode) => ussActions.refreshUSSInTree(node, ussFileProvider))
     );
     context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.refreshDirectory", (node, nodeList) => {
+        vscode.commands.registerCommand("zowe.uss.refreshDirectory", async (node, nodeList) => {
             let selectedNodes = getSelectedNodeList(node, nodeList) as IZoweUSSTreeNode[];
             selectedNodes = selectedNodes.filter((x) => contextuals.isUssDirectory(x));
             for (const item of selectedNodes) {
-                ussActions.refreshDirectory(item, ussFileProvider);
+                await ussActions.refreshDirectory(item, ussFileProvider);
             }
         })
     );
@@ -84,13 +84,13 @@ export async function initUSSProvider(context: vscode.ExtensionContext): Promise
         vscode.commands.registerCommand("zowe.uss.ZoweUSSNode.open", (node: IZoweUSSTreeNode): void => node.openUSS(false, true, ussFileProvider))
     );
     context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.removeSession", (node: IZoweUSSTreeNode, nodeList) => {
+        vscode.commands.registerCommand("zowe.uss.removeSession", async (node: IZoweUSSTreeNode, nodeList) => {
             let selectedNodes = getSelectedNodeList(node, nodeList);
             selectedNodes = selectedNodes.filter((element) => contextuals.isUssSession(element));
             for (const item of selectedNodes) {
                 ussFileProvider.deleteSession(item);
             }
-            TreeViewUtils.fixVsCodeMultiSelect(ussFileProvider);
+            await TreeViewUtils.fixVsCodeMultiSelect(ussFileProvider);
         })
     );
     context.subscriptions.push(
@@ -151,9 +151,17 @@ export async function initUSSProvider(context: vscode.ExtensionContext): Promise
         vscode.commands.registerCommand("zowe.uss.removeFavProfile", (node): void => ussFileProvider.removeFavProfile(node.label, true))
     );
     context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.disableValidation", (node) => Profiles.getInstance().disableValidation(node))
+        vscode.commands.registerCommand("zowe.uss.disableValidation", (node) => {
+            Profiles.getInstance().disableValidation(node);
+            ussFileProvider.refreshElement(node);
+        })
     );
-    context.subscriptions.push(vscode.commands.registerCommand("zowe.uss.enableValidation", (node) => Profiles.getInstance().enableValidation(node)));
+    context.subscriptions.push(
+        vscode.commands.registerCommand("zowe.uss.enableValidation", (node) => {
+            Profiles.getInstance().enableValidation(node);
+            ussFileProvider.refreshElement(node);
+        })
+    );
     context.subscriptions.push(vscode.commands.registerCommand("zowe.uss.ssoLogin", (node: IZoweTreeNode): void => ussFileProvider.ssoLogin(node)));
     context.subscriptions.push(vscode.commands.registerCommand("zowe.uss.ssoLogout", (node: IZoweTreeNode): void => ussFileProvider.ssoLogout(node)));
     context.subscriptions.push(
@@ -162,7 +170,7 @@ export async function initUSSProvider(context: vscode.ExtensionContext): Promise
                 await ussFileProvider.copying;
             }
 
-            ussActions.pasteUss(ussFileProvider, node);
+            await ussActions.pasteUss(ussFileProvider, node);
         })
     );
     context.subscriptions.push(