+++ packages/zowe-explorer/src/utils/ProfilesUtils.ts
@@ -139,7 +139,6 @@ export const syncSessionNode =
     (getSessionForProfile: SessionForProfile) =>
     (sessionNode: IZoweTreeNode): void => {
         ZoweLogger.trace("ProfilesUtils.syncSessionNode called.");
-        sessionNode.collapsibleState = vscode.TreeItemCollapsibleState.Collapsed;
 
         const profileType = sessionNode.getProfile().type;
         const profileName = sessionNode.getProfileName();
@@ -306,9 +305,7 @@ export class ProfilesUtils {
         if (!fs.existsSync(settingsPath)) {
             fs.mkdirSync(settingsPath);
         }
-        if (!fs.existsSync(path.join(settingsPath, "imperative.json"))) {
-            ProfilesUtils.writeOverridesFile();
-        }
+        ProfilesUtils.writeOverridesFile();
         // If not using team config, ensure that the ~/.zowe/profiles directory
         // exists with appropriate types within
         if (!imperative.ImperativeConfig.instance.config?.exists) {
@@ -322,43 +319,53 @@ export class ProfilesUtils {
 
     public static writeOverridesFile(): void {
         ZoweLogger.trace("ProfilesUtils.writeOverridesFile called.");
-        let fd: number;
-        let fileContent: string;
+        const defaultImperativeJson = { overrides: { CredentialManager: globals.ZOWE_CLI_SCM } };
         const settingsFile = path.join(getZoweDir(), "settings", "imperative.json");
+        let fileContent: string;
         try {
-            fd = fs.openSync(settingsFile, "r+");
-            fileContent = fs.readFileSync(fd, "utf-8");
-        } catch {
-            // If reading the file failed because it does not exist, then create it
-            // This should never fail, unless file system is read-only or the file
-            // was created by another process after first openSync call
-            fd = fs.openSync(settingsFile, "wx");
+            fileContent = fs.readFileSync(settingsFile, { encoding: "utf-8" });
+        } catch (error) {
+            ZoweLogger.debug(localize("writeOverridesFile.readFile.error", "Reading imperative.json failed. Will try to create file."));
         }
-        try {
-            let settings: any;
-            if (fileContent) {
-                try {
-                    settings = JSON.parse(fileContent);
-                } catch (err) {
-                    if (err instanceof Error) {
-                        throw new Error(
-                            localize("writeOverridesFile.jsonParseError", "Failed to parse JSON file {0}:", settingsFile) + " " + err.message
-                        );
-                    }
+        let settings: any;
+        if (fileContent) {
+            try {
+                settings = JSON.parse(fileContent);
+                ZoweLogger.debug(localize("writeOverridesFile.readFile", "Reading imperative.json Credential Manager.\n {0}", fileContent));
+            } catch (err) {
+                if (err instanceof Error) {
+                    const errorMsg = localize(
+                        "writeOverridesFile.jsonParse.error",
+                        "Failed to parse JSON file {0}. Will try to re-create the file.",
+                        settingsFile
+                    );
+                    ZoweLogger.error(errorMsg);
+                    ZoweLogger.debug(fileContent.toString());
+                    settings = { ...defaultImperativeJson };
+                    fs.writeFileSync(settingsFile, JSON.stringify(settings, null, 2), {
+                        encoding: "utf-8",
+                        flag: "w",
+                    });
                 }
-                if (settings && settings?.overrides && settings?.overrides?.CredentialManager !== globals.PROFILE_SECURITY) {
-                    settings.overrides.CredentialManager = globals.PROFILE_SECURITY;
-                } else {
-                    return;
-                }
-            } else {
-                settings = { overrides: { CredentialManager: globals.PROFILE_SECURITY } };
             }
-            fileContent = JSON.stringify(settings, null, 2);
-            fs.writeFileSync(fd, fileContent, "utf-8");
-        } finally {
-            fs.closeSync(fd);
+            if (settings?.overrides?.CredentialManager === globals.PROFILE_SECURITY) {
+                return;
+            }
+            if (!settings?.overrides?.CredentialManager) {
+                settings = { ...defaultImperativeJson, ...settings };
+            }
+        } else {
+            settings = { ...defaultImperativeJson };
         }
+        settings.overrides.CredentialManager = globals.PROFILE_SECURITY;
+        const newData = JSON.stringify(settings, null, 2);
+        ZoweLogger.debug(
+            localize("writeOverridesFile.updateFile", "Updating imperative.json Credential Manager to {0}.\n{1}", globals.PROFILE_SECURITY, newData)
+        );
+        fs.writeFileSync(settingsFile, newData, {
+            encoding: "utf-8",
+            flag: "w",
+        });
     }
 
     public static async initializeZoweProfiles(): Promise<void> {
@@ -375,7 +382,7 @@ export class ProfilesUtils {
             ZoweLogger.info(localize("initializeZoweProfiles.success", "Zowe Profiles initialized successfully."));
         } catch (err) {
             if (err instanceof imperative.ImperativeError) {
-                errorHandling(err, undefined, err.mDetails.causeErrors);
+                await errorHandling(err, undefined, err.mDetails.causeErrors);
             } else {
                 ZoweLogger.error(err);
                 ZoweExplorerExtender.showZoweConfigError(err.message);