+++ scripts/d2uChangelog.js
@@ -9,8 +9,5 @@
  *
  */
 
-import { createSystemTestEnvironment } from "./systemTestEnv";
-
-(async () => {
-    await createSystemTestEnvironment();
-})();
+const dos2unix = require("ssp-dos2unix").dos2unix;
+let converted = dos2unix("CHANGELOG.md", { feedback: true, writable: true });