+++ packages/zowe-explorer/CHANGELOG.md
@@ -12,6 +12,25 @@ All notable changes to the "vscode-extension-for-zowe" extension will be documen
 
 ### Bug fixes
 
+- Fixed issue where the "Disable Validation for Profile" context menu option did not update to "Enable Validation for Profile" after use. [#1897](https://github.com/zowe/vscode-extension-for-zowe/issues/1897)
+- Removed "/" characters in `path.join()` calls [#2172](https://github.com/zowe/vscode-extension-for-zowe/issues/2172)
+
+## `2.9.0` IN PROGRESS
+
+### New Features and Enhancements
+
+- Added "Cancel Job" feature for job nodes in Jobs tree view. [#2251](https://github.com/zowe/vscode-extension-for-zowe/issues/2251)
+
+## `2.8.1`
+
+### Bug fixes
+
+- Fixed an issue with updating imperative.json file's Credential Manager value. [#2289](https://github.com/zowe/vscode-extension-for-zowe/issues/2289)
+- Fixed an issue with "Zowe Explorer: Poll Content in Active Editor" keybind interfering with debug capabilities in VScode. The keybind to poll JES Spool file content will require the spool file to be active in the text editor. [#2285](https://github.com/zowe/vscode-extension-for-zowe/issues/2285)
+- Updated linter rules and addressed linter errors. [#2291](https://github.com/zowe/vscode-extension-for-zowe/issues/2291)
+- Fixed an issue with `zowe.settings.version` setting being updated with incorrect type. [#2166](https://github.com/zowe/vscode-extension-for-zowe/issues/2166)
+- Updated dependencies for security audits.
+
 ## `2.8.0`
 
 ### New features and enhancements
@@ -35,7 +54,9 @@ All notable changes to the "vscode-extension-for-zowe" extension will be documen
 - Fixed redundant text in error messages that included the same error details twice.
 - Fixed issue where a spool file would open a duplicate tab when clicked between updates. [#1952](https://github.com/zowe/vscode-extension-for-zowe/issues/1952)
 - Fixed issue where a job search query would not expand the session node after it has been filtered.
+- Fixed issue where refreshing views collapsed the respective trees. [#2215](https://github.com/zowe/vscode-extension-for-zowe/issues/2215)
 - Fixed error message when no data sets found that match pattern.
+- Fixed secure credential storage not possible to enable in Theia.
 
 ## `2.7.0`